changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11764 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/SerialPorts/ORSerialPortList.h
@@ -34,7 +34,7 @@
 - (void) updatePortList;
 
 - (NSUInteger) count;
-- (ORSerialPort*) objectAtIndex:(unsigned)index;
+- (ORSerialPort*) objectAtIndex:(NSUInteger)index;
 - (ORSerialPort*) objectWithName:(NSString*) name;
 - (NSInteger) indexOfObjectWithName:(NSString*) name;
 - (NSArray*) portList;
